backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
celery==5.5.3
django-environ==0.11.0
psycopg[binary]==3.2.12
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.4.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0
ruff==0.8.5
mypy==1.18.2
django-stubs[compatible-mypy]==5.2.7
djangorestframework-stubs==3.16.1
types-redis==4.6.0.20241004
